Api Server/requirements.txt
fastapi
google_cloud_aiplatform
pydantic
python-dotenv
uvicorn
